Revert "feat(BELAT-528): :art: new screens for module mange projects"
This reverts commit 8245d02aaeca4c628ab497300b3548920782b21a.
