+++ src/app/clients/clients-view/custom-dialogs/upload-document-dialog/upload-document-dialog.component.ts
@@ -30,10 +30,11 @@ export class UploadDocumentDialogComponent implements OnInit {
     private formBuilder: UntypedFormBuilder,
     @Inject(MAT_DIALOG_DATA) public data: any
   ) {
+    console.log(data);
     this.documentIdentifier = data.documentIdentifier;
     this.entityType = data.entityType;
-    this.documentClassOptions = data?.documentClassOptions?.codeValues || [];
-    this.documentTypeOptions = data?.documentTypeOptions?.codeValues || [];
+    // this.documentClassOptions = data.documentClassOptions.codeValues;
+    // this.documentTypeOptions = data.documentTypeOptions.codeValues;
   }
 
   ngOnInit() {
