--- src/app/organization/manage-projects/view-investment-project/investment-project-documents-tab/investment-project-documents-tab.component.scss
@@ -1,16 +1,0 @@
-table {
-  width: 100%;
-  margin-top: 3%;
-
-  .document-action-button {
-    min-width: 26px;
-    padding: 0 6px;
-    margin: 4px;
-    line-height: 25px;
-  }
-}
-
-.tab-container {
-  padding: 1%;
-  margin: 1%;
-}
