+++ src/app/organization/manage-projects/view-investment-project/investment-project-general-tab/investment-project-general-tab.component.scss
@@ -8,8 +8,3 @@
     }
   }
 }
-
-.tab-container {
-  padding: 1%;
-  margin: 1%;
-}
