+++ src/app/organization/manage-projects/view-investment-project/investment-project-image-tab/investment-project-image-tab.component.html
@@ -1,121 +1,75 @@
-<div class="tab-container mat-typography">
-  <div fxLayout="row" fxLayoutAlign="end" fxLayoutGap="2%" fxLayout.lt-md="column" class="container m-b-20 m-r-0">
-    <button mat-raised-button color="primary" [routerLink]="['../edit']" *mifosxHasPermission="'UPDATE_FUND'">
-      <fa-icon icon="edit"></fa-icon>
-      {{ 'labels.buttons.Edit' | translate }}
-    </button>
-  </div>
-  <section class="container">
-    <mat-card class="mainCard">
-      <mat-card-content fxLayout="column" class="m-t-10">
-        <div class="content m-b-20">
-          <div fxLayout="row wrap" class="m-b-15">
-            <div fxFlex="30%" class="mat-body-strong">
-              {{ 'labels.inputs.custom.Institution Description' | translate }}
-            </div>
-            <div fxFlex="70%">
-              {{ shorten(projectData.institutionDescription, 100) }}
-            </div>
-          </div>
+<div class="container m-b-30">
+  <mat-card>
+    <mat-card-title class="text-center">{{ 'labels.heading.Cover Image' | translate }}</mat-card-title>
+    <mat-divider></mat-divider>
 
-          <div fxLayout="row wrap" class="m-b-15">
-            <div fxFlex="30%" class="mat-body-strong">
-              {{ 'labels.inputs.custom.Team Description' | translate }}
-            </div>
-            <div fxFlex="70%">
-              {{ shorten(projectData.teamDescription, 100) }}
-            </div>
-          </div>
-
-          <div fxLayout="row wrap" class="m-b-15">
-            <div fxFlex="30%" class="mat-body-strong">
-              {{ 'labels.inputs.custom.Financial Description' | translate }}
-            </div>
-            <div fxFlex="70%">
-              {{ shorten(projectData.financingDescription, 100) }}
-            </div>
-          </div>
-
-          <div fxLayout="row wrap" class="m-b-15">
-            <div fxFlex="30%" class="mat-body-strong">
-              {{ 'labels.inputs.name' | translate }}
-            </div>
-            <div fxFlex="70%">
-              {{ projectData.name }}
-            </div>
-          </div>
-
-          <div fxLayout="row wrap" class="m-b-15">
-            <div fxFlex="30%" class="mat-body-strong">
-              {{ 'labels.inputs.Portal Position' | translate }}
-            </div>
-            <div fxFlex="70%">
-              {{ projectData.position }}
-            </div>
-          </div>
-
-          <div fxLayout="row wrap" class="m-b-15">
-            <div fxFlex="30%" class="mat-body-strong">
-              {{ 'labels.inputs.Visible On Portal' | translate }}
-            </div>
-            <div fxFlex="70%">
-              {{ projectData.isActive | yesNo | translate }}
-            </div>
-          </div>
-        </div>
-
-        <h3 fxFlex="90%" class="mat-body-strong">{{ 'labels.heading.Portrait image' | translate }}</h3>
-
-        <div class="m-b-15">
-          <mat-card class="thumbnail-card">
-            <img
-              mat-card-image
-              *ngIf="coverImage"
-              [src]="getImagePath(coverImage.location)"
-              alt="Image"
-              class="project-img"
-            />
-            <button mat-icon-button *ngIf="coverImage" class="delete-btn" (click)="deleteDocument(coverImage.id)">
-              <mat-icon>close</mat-icon>
-            </button>
-            <button *ngIf="!coverImage" class="addImageButton" (click)="uploadCoverDocument()">
-              <fa-icon icon="plus" class="m-r-5"></fa-icon>
-            </button>
-          </mat-card>
-        </div>
-        <h3 fxFlex="90%" class="mat-body-strong">{{ 'labels.heading.Carrousel images' | translate }}</h3>
-        <div fxLayout="row" fxLayoutWrap class="thumbnail-gallery">
-          <ng-container *ngFor="let image of imageData">
-            <div class="thumbnail-wrapper" *ngIf="image.description !== 'Cover'">
-              <mat-card class="thumbnail-card">
-                <img mat-card-image [src]="getImagePath(image.location)" alt="Image" class="project-img" />
-
-                <button mat-icon-button class="delete-btn" (click)="deleteDocument(image.id)">
-                  <mat-icon>close</mat-icon>
-                </button>
-              </mat-card>
-
-              <mat-form-field appearance="fill" class="order-field m-t-10">
-                <mat-label>{{ 'labels.inputs.Order' | translate }}</mat-label>
-                <input matInput type="number" [(ngModel)]="image.description" (ngModelChange)="image.modified = true" />
-
-                <button
-                  mat-icon-button
-                  matSuffix
-                  color="accent"
-                  (click)="updateImageOrder(image)"
-                  [disabled]="!image.modified"
-                >
-                  <mat-icon>save</mat-icon>
-                </button>
-              </mat-form-field>
-            </div>
-          </ng-container>
-          <button class="addImageButton" (click)="uploadDocument(null)">
-            <fa-icon icon="plus" class="m-r-5"></fa-icon>
+    <div
+      fxLayout="row"
+      fxLayoutAlign="end center"
+      fxLayoutGap="10px"
+      fxLayout.lt-md="column"
+      class="m-b-20"
+      *ngIf="!coverImage"
+    >
+      <button mat-raised-button color="primary" (click)="uploadCoverDocument()">
+        <fa-icon icon="plus" class="m-r-5"></fa-icon>
+        {{ 'labels.buttons.Add' | translate }}
+      </button>
+    </div>
+    <mat-card-content fxLayout="column" fxLayoutAlign="center center" class="m-t-10">
+      <div *ngIf="coverImage">
+        <mat-card class="thumbnail-card">
+          <img mat-card-image [src]="getImagePath(coverImage.location)" alt="Image" class="project-img" />
+          <button mat-icon-button class="delete-btn" (click)="deleteDocument(coverImage.id)">
+            <mat-icon>close</mat-icon>
           </button>
-        </div>
-      </mat-card-content>
-    </mat-card>
-  </section>
+        </mat-card>
+      </div>
+    </mat-card-content>
+  </mat-card>
+</div>
+
+<div class="container m-b-30">
+  <mat-card>
+    <mat-card-title class="text-center">{{ 'labels.heading.Other Images' | translate }}</mat-card-title>
+    <mat-divider></mat-divider>
+
+    <mat-card-content>
+      <div fxLayout="row" fxLayoutAlign="end center" fxLayoutGap="10px" fxLayout.lt-md="column" class="m-b-20">
+        <button mat-raised-button color="primary" (click)="uploadDocument(null)">
+          <fa-icon icon="plus" class="m-r-5"></fa-icon>
+          {{ 'labels.buttons.Add' | translate }}
+        </button>
+      </div>
+
+      <div fxLayout="row" fxLayoutGap="16px" fxLayoutAlign="center start" fxLayoutWrap class="thumbnail-gallery">
+        <ng-container *ngFor="let image of imageData">
+          <div class="thumbnail-wrapper" *ngIf="image.description !== 'Cover'">
+            <mat-card class="thumbnail-card">
+              <img mat-card-image [src]="getImagePath(image.location)" alt="Image" class="project-img" />
+
+              <button mat-icon-button class="delete-btn" (click)="deleteDocument(image.id)">
+                <mat-icon>close</mat-icon>
+              </button>
+            </mat-card>
+
+            <mat-form-field appearance="fill" class="order-field m-t-10">
+              <mat-label>{{ 'labels.inputs.Order' | translate }}</mat-label>
+              <input matInput type="number" [(ngModel)]="image.description" (ngModelChange)="image.modified = true" />
+
+              <button
+                mat-icon-button
+                matSuffix
+                color="accent"
+                (click)="updateImageOrder(image)"
+                [disabled]="!image.modified"
+              >
+                <mat-icon>save</mat-icon>
+              </button>
+            </mat-form-field>
+          </div>
+        </ng-container>
+      </div>
+    </mat-card-content>
+  </mat-card>
 </div>
