+++ src/app/organization/manage-projects/view-investment-project/investment-project-image-tab/investment-project-image-tab.component.scss
@@ -58,38 +58,3 @@
     border-radius: 4px;
   }
 }
-
-.tab-container {
-  padding: 1%;
-  margin: 1%;
-}
-
-.imagesSection {
-  max-width: 65%;
-}
-
-.addImageButton {
-  background-color: transparent;
-  font-weight: 300;
-  font-size: 2.2rem;
-  outline: none;
-  align-self: center;
-  border: none;
-  cursor: pointer;
-  padding-left: 0.3em;
-}
-
-.thumbnail-gallery {
-  overflow: auto;
-  min-width: 0;
-  gap: 10px;
-  scroll-snap-align: start;
-
-  &::-webkit-scrollbar {
-    height: 2px;
-  }
-}
-
-.tab-container .mainCard {
-  padding: 30px;
-}
