--- src/app/organization/manage-projects/view-investment-project/investment-project-impact-tab/investment-project-impact-tab.component.scss
@@ -1,8 +1,0 @@
-.mainCard {
-  padding: 30px;
-}
-
-.tab-container {
-  padding: 1%;
-  margin: 1%;
-}
