--- src/app/organization/manage-projects/view-investment-project/investment-project-notes-tab/investment-project-notes-tab.component.scss
@@ -1,4 +1,0 @@
-.tab-container {
-  padding: 1%;
-  margin: 1%;
-}
