--- src/app/organization/manage-projects/view-investment-project/investment-project-simulate-tab/investment-project-simulate-tab.component.scss
@@ -1,4 +1,0 @@
-.tab-container {
-  padding: 1%;
-  margin: 1%;
-}
